Fixes IncCheck and DecCheck
They were not actually modifying the variable on the top of
stack, but they were creating another variable! In all the other
cases they worked perfectly fine.

Configures sane logging to a log file.

Don't add OPs in the table if not strictly necessary! We must check
they are working correctly before moving on :)